hotfix: change bounds for security status classification
The system of Sakhti has a true security status of 0.454... as reported by ESI.
Under the current bounds, this was being classed as low-security space.
EVE Online treats the system as high-security space, indicating a rounding of truesec.
We therefore change the bound to reflect this rounding and correctly display the system's security class.